increased id sizes

+++ src/stat_service/db/init/create_like_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS likes
 (
-    post_id Int32,
-    author_id Int32,
-    user_id Int32,
+    post_id Int64,
+    author_id Int64,
+    user_id Int64,
     timestamp DateTime,
 ) ENGINE = ReplacingMergeTree
 PRIMARY KEY (post_id, user_id)
